# api/config/openapi.yml
openapi: 3.0.0
info:
  title: Image Upload API
  version: "2.0.0"

paths:
  /analyze:
    post:
      summary: Upload an image and run analysis.
      operationId: api.app.upload_image_and_process

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
            encoding:
              image:
                contentType: image/*

      responses:
        '200':
          description: File processed successfully.
          content:
            application/json:
              schema:
                type: object
                required: [schema_version, job_id, pipeline, qc, objects]
                properties:
                  schema_version: { type: string }
                  job_id: { type: string }
                  timestamp: { type: string }
                  pipeline:
                    type: object
                    properties:
                      name: { type: string }
                      version: { type: string }
                  qc: { type: object }
                  traits_emitted:
                    type: array
                    items: { type: string }
                  derived_images:
                    type: array
                    items:
                      type: object
                      properties:
                        role: { type: string }
                        filename: { type: string }
                        url: { type: string }
                  objects:
                    type: array
                    items:
                      type: object
                      required: [object_id, traits]
                      properties:
                        object_id: { type: string }
                        source_label: { type: string }
                        bbox: { type: object }
                        qc: { type: object }
                        traits:
                          type: object
                          additionalProperties:
                            type: object
                            properties:
                              value: { nullable: true }
                              unit: { nullable: true }

        '400':
          description: Missing image or wrong type.